Revert "Landing Fixed (#26)"
This reverts commit 559e5465fbf8fdc99e5ace269b31df0544eaf1bc.
