+++ src/Component/HomeComponents/Landing/Landing.css
@@ -1,64 +1,58 @@
-.landing {
-  width: 100%;
-  position: relative;
+* {
+    margin: 0;
+    padding: 0;
 }
 
-.landing__text {
-  position: absolute;
-  top: 30%;
-  left: 50%;
-  transform: translateX(-50%);
+.landing{
+    background-color: #1e022c;
+    text-align: center;
 }
 
-.landing__text--font {
-  font-weight: 500;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: normal;
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-  position: relative;
-  text-align: center;
+.kreiva{
+    background: linear-gradient(to bottom, #ffffff, rgba(255, 255, 255, 0.51), rgba(255, 255, 255, 0));
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+
+    font-size: 9.5em;
+    z-index: 20;
+
+    font-family: Noir;
+    padding-top: 19%;
 }
 
-.landing__text--section {
-  background-image: linear-gradient(
-    to bottom,
-    #ffffff,
-    rgba(255, 255, 255, 0.51),
-    rgba(255, 255, 255, 0)
-  );
-  font-size: 200px;
-}
-.landing__text--sub-section {
-  background-image: linear-gradient(
-    to bottom,
-    #ffffff,
-    rgba(139, 139, 139, 0.55),
-    rgba(0, 0, 0, 0)
-  );
-  font-size: 100px;
-  transform: translateY(-75%);
-}
-@media screen and (max-width: 1280px) {
-  .landing__text {
-  }
-  .landing__text--section {
-    font-size: 150px;
-  }
-  .landing__text--sub-section {
-    font-size: 77px;
-    transform: translateY(-60%);
-  }
+.virtual{
+    background: linear-gradient(to bottom, #821169, rgba(130, 17, 105, 0));
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+
+  
+    font-size: 2em;
+    padding: 1%;
 }
 
-@media screen and (max-width: 800px) {
-  .landing__text--section {
-    font-size: 77px;
-  }
-  .landing__text--sub-section {
-    font-size: 48px;
-  }
+.parent{
+    position: relative;
 }
+
+.a{
+    font-size: 4.5em;
+    padding: 0px;
+    margin: 0px;
+    
+    position: absolute;
+    top: 77%;
+    justify-content: center;
+    left: 43%;
+
+    background-image: linear-gradient(to bottom, #ffffff, rgba(139, 139, 139, 0.55), rgba(0, 0, 0, 0));
+}
+
+
+.wave{
+    position: absolute;
+    left: 0;
+    z-index: 0;
+    top: -2;
+    width: 100%;
+}
+
